testplan-agent/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
groq==0.4.2
ollama==0.1.6
pdfplumber==0.10.0
python-docx==1.1.0
reportlab==4.0.9
markdown==3.5.2
requests==2.31.0
httpx==0.25.2
